pyproject.toml
[tool.poetry]
name = "codegen"
version = "0.1.0"
description = "Code Generator"
authors = ["radzek15 <radzek15@wp.pl>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.6"
psycopg2-binary = "^2.9.9"
ruff = "^0.1.1"
pre-commit = "^3.5.0"
isort = "^5.12.0"
pyupgrade = "^3.15.0"
djlint = "^1.34.0"
django-environ = "^0.11.2"
argon2-cffi = "^23.1.0"
djangorestframework = "^3.14.0"
django-filter = "^23.3"
djangorestframework-simplejwt = "^5.3.0"
dj-rest-auth = {extras = ["with-social"], version = "^5.0.1"}
django-cors-headers = "^4.3.0"
djangorestframework-yaml = "^2.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
ignore = [
    "E712",  # Allow using if x == False, as it's not always equivalent to if x.
    "E501",  # Supress line-too-long warnings: trust black's judgement on this one.
    "A003",  # Allow shawoding class attribute: django uses id.
    "B904",  # Allow unchained exceptions: it's fine to raise 404 in django.
    "F401",  # Unused import
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["E402"]

[tool.isort]
profile = "django"
combine_as_imports = true
include_trailing_comma = true
line_length = 120
multi_line_output = 3

[tool.djlint]
profile="django"
ignore = "H031"
